refactor(logging): split runtime roles
